+++ darts/models/__init__.py
@@ -89,10 +89,11 @@ except ModuleNotFoundError:
 
 try:
     from darts.models.forecasting.croston import Croston
-    from darts.models.forecasting.sf_models import (
-        StatsForecastAutoARIMA,
-        StatsForecastETS,
-    )
+    from darts.models.forecasting.sf_auto_arima import StatsForecastAutoARIMA
+    from darts.models.forecasting.sf_auto_ces import StatsForecastAutoCES
+    from darts.models.forecasting.sf_auto_theta import StatsForecastAutoTheta
+    from darts.models.forecasting.sf_ets import StatsForecastETS
+
 except ImportError:
     logger.warning(
         "The statsforecast module could not be imported. "
