+++ darts/tests/models/forecasting/test_local_forecasting_models.py
@@ -31,6 +31,7 @@ from darts.models import (
     RandomForest,
     RegressionModel,
     StatsForecastAutoARIMA,
+    StatsForecastAutoTheta,
     StatsForecastETS,
     Theta,
 )
@@ -51,6 +52,11 @@ models = [
     (ARIMA(12, 2, 1), 5.2),
     (ARIMA(1, 1, 1), 24),
     (StatsForecastAutoARIMA(season_length=12), 4.6),
+    (
+        StatsForecastAutoTheta(season_length=12, decomposition_type="multiplicative"),
+        5.5,
+    ),
+    (StatsForecastAutoTheta(season_length=12, decomposition_type="additive"), 7.9),
     (StatsForecastETS(season_length=12, model="AAZ"), 4.1),
     (Croston(version="classic"), 23),
     (Croston(version="tsb", alpha_d=0.1, alpha_p=0.1), 23),
