+++ requirements/core.txt
@@ -13,7 +13,7 @@ requests>=2.22.0
 scikit-learn>=1.0.1
 scipy>=1.3.2
 shap>=0.40.0
-statsforecast>=1.0.0
+statsforecast>=1.4.0
 statsmodels>=0.13.0
 tbats>=1.1.0
 tqdm>=4.60.0
